+++ src/main/kotlin/com/github/naoyukik/intellijplugingithubnotifications/application/ApiClientWorkflow.kt
@@ -1,6 +1,7 @@
 package com.github.naoyukik.intellijplugingithubnotifications.application
 
-import com.github.naoyukik.intellijplugingithubnotifications.application.dto.TableDataDto
+import com.github.naoyukik.intellijplugingithubnotifications.application.dto.GitHubNotificationDto
+import com.github.naoyukik.intellijplugingithubnotifications.application.dto.NotificationDetailDto
 import com.github.naoyukik.intellijplugingithubnotifications.domain.GitHubNotificationRepository
 import com.github.naoyukik.intellijplugingithubnotifications.domain.SettingStateRepository
 import com.github.naoyukik.intellijplugingithubnotifications.domain.model.GitHubNotification
@@ -8,14 +9,15 @@ import com.github.naoyukik.intellijplugingithubnotifications.domain.model.GitHub
 import com.github.naoyukik.intellijplugingithubnotifications.domain.model.NotificationDetailResponse.NotificationDetail
 import com.github.naoyukik.intellijplugingithubnotifications.domain.model.NotificationDetailResponse.NotificationDetailError
 import com.github.naoyukik.intellijplugingithubnotifications.domain.model.SettingState
-import com.intellij.openapi.util.IconLoader
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import java.net.URI
-import java.net.URL
 import java.time.ZonedDateTime
-import javax.swing.Icon
+import com.github.naoyukik.intellijplugingithubnotifications.application.dto.GitHubNotificationDto.Repository as DtoRepository
+import com.github.naoyukik.intellijplugingithubnotifications.application.dto.GitHubNotificationDto.Subject as DtoSubject
+import com.github.naoyukik.intellijplugingithubnotifications.application.dto.GitHubNotificationDto.SubjectType as DtoSubjectType
+import com.github.naoyukik.intellijplugingithubnotifications.application.dto.NotificationDetailDto.NotificationDetail.RequestedReviewers as DtoRequestedReviewers
 
 class ApiClientWorkflow(
     private val repository: GitHubNotificationRepository,
@@ -24,66 +26,7 @@ class ApiClientWorkflow(
 ) {
     private var latestFetchTime: ZonedDateTime? = null
 
-    companion object {
-        private val TYPE_TO_PATH = mapOf(
-            "PullRequest" to "pull",
-            "Issue" to "issues",
-            "Release" to "releases",
-        )
-
-        private val issuesOpened = IconLoader.getIcon(
-            "com/github/naoyukik/intellijplugingithubnotifications/icons/issue-opened-16.svg",
-            this::class.java.classLoader,
-        )
-
-        private val issuesClosed = IconLoader.getIcon(
-            "com/github/naoyukik/intellijplugingithubnotifications/icons/issue-closed-16.svg",
-            this::class.java.classLoader,
-        )
-
-        private val pullRequestsOpen = IconLoader.getIcon(
-            "/com/github/naoyukik/intellijplugingithubnotifications/icons/git-pull-request-16.svg",
-            this::class.java.classLoader,
-        )
-
-        private val pullRequestsMerged = IconLoader.getIcon(
-            "/com/github/naoyukik/intellijplugingithubnotifications/icons/git-merge-16.svg",
-            this::class.java.classLoader,
-        )
-
-        private val pullRequestsClosed = IconLoader.getIcon(
-            "/com/github/naoyukik/intellijplugingithubnotifications/icons/git-pull-request-closed-16.svg",
-            this::class.java.classLoader,
-        )
-
-        private val pullRequestsDraft = IconLoader.getIcon(
-            "/com/github/naoyukik/intellijplugingithubnotifications/icons/git-pull-request-draft-16.svg",
-            this::class.java.classLoader,
-        )
-
-        private val release = IconLoader.getIcon(
-            "/com/github/naoyukik/intellijplugingithubnotifications/icons/tag-16.svg",
-            this::class.java.classLoader,
-        )
-
-        private val unread = IconLoader.getIcon(
-            "/com/github/naoyukik/intellijplugingithubnotifications/icons/unread-16.svg",
-            this::class.java.classLoader,
-        )
-
-        private val TYPE_TO_EMOJI: Map<String, Icon> = mapOf(
-            "PullRequestOpen" to pullRequestsOpen,
-            "PullRequestMerged" to pullRequestsMerged,
-            "PullRequestClosed" to pullRequestsClosed,
-            "PullRequestDraft" to pullRequestsDraft,
-            "IssueOpen" to issuesOpened,
-            "IssueClosed" to issuesClosed,
-            "Release" to release,
-            "Unread" to unread,
-        )
-    }
-
-    suspend fun fetchNotifications(): List<TableDataDto> = withContext(dispatcher) {
+    suspend fun fetchNotifications(): List<GitHubNotificationDto> = withContext(dispatcher) {
         val settingState = settingStateRepository.loadSettingState()
         val ghCliPath = settingState.ghCliPath
         val includingRead = settingState.includingRead
@@ -119,7 +62,7 @@ class ApiClientWorkflow(
             updatedNotification
         }
 
-        updatedNotifications.toTableDataDto()
+        updatedNotifications.toGitHubNotificationDto()
     }
 
     private fun hasNewNotificationsSinceLastCheck(
@@ -169,77 +112,34 @@ class ApiClientWorkflow(
         }
     }
 
-    private fun List<GitHubNotification>.toTableDataDto(): List<TableDataDto> {
+    private fun List<GitHubNotification>.toGitHubNotificationDto(): List<GitHubNotificationDto> {
         return this.map {
-            val issueNumber = getIssueNumber(it.subject.url)
-            TableDataDto(
-                title = it.subject.title,
-                fullName = apiUrlToRepositoryIssueNumberConverter(
-                    repositoryFullName = it.repository.fullName,
-                    issueNumber = issueNumber,
-                    type = it.subject.type,
-                ),
-                htmlUrl = apiUrlToHtmlUrlConverter(
-                    htmlUrl = it.repository.htmlUrl,
-                    issueNumber = issueNumber,
-                    type = it.subject.type,
-                ),
+            GitHubNotificationDto(
+                id = it.id,
                 reason = it.reason,
                 updatedAt = it.updatedAt,
-                typeEmoji = apiUrlToEmojiConverter(it),
-                reviewers = it.detail?.requestedReviewers?.map { reviewer -> reviewer.login } ?: emptyList(),
-                unreadEmoji = it.unread.takeIf { it }?.let { TYPE_TO_EMOJI["Unread"] },
+                unread = it.unread,
+                subject = DtoSubject(
+                    title = it.subject.title,
+                    url = it.subject.url,
+                    type = DtoSubjectType.valueOf(
+                        it.subject.type.name,
+                    ),
+                ),
+                repository = DtoRepository(
+                    fullName = it.repository.fullName,
+                    htmlUrl = it.repository.htmlUrl,
+                ),
+                detail = NotificationDetailDto.NotificationDetail(
+                    state = it.detail?.state.toString(),
+                    merged = it.detail?.merged == true,
+                    draft = it.detail?.draft == true,
+                    htmlUrl = it.detail?.htmlUrl.toString(),
+                    requestedReviewers = it.detail?.requestedReviewers?.map { reviewer ->
+                        DtoRequestedReviewers(login = reviewer.login)
+                    } ?: emptyList(),
+                ),
             )
         }
     }
-
-    private fun apiUrlToRepositoryIssueNumberConverter(
-        repositoryFullName: String,
-        issueNumber: String,
-        type: SubjectType,
-    ): String {
-        return TYPE_TO_PATH[type.name]?.let { typeToPath ->
-            when (typeToPath) {
-                "issues", "pull" -> "$repositoryFullName #$issueNumber"
-                "releases" -> "$issueNumber in $repositoryFullName"
-                else -> "$repositoryFullName #$issueNumber"
-            }
-        } ?: "$repositoryFullName #$issueNumber"
-    }
-
-    private fun apiUrlToHtmlUrlConverter(htmlUrl: String, issueNumber: String, type: SubjectType): URL? {
-        return TYPE_TO_PATH[type.name]?.let { typeToPath ->
-            when (typeToPath) {
-                "issues", "pull" -> URI("$htmlUrl/$typeToPath/$issueNumber").toURL()
-                "releases" -> URI("$htmlUrl/$typeToPath/tag/$issueNumber").toURL()
-                else -> null
-            }
-        }
-    }
-
-    @Suppress("ComplexMethod")
-    private fun apiUrlToEmojiConverter(notification: GitHubNotification): Icon? {
-        return when (val type = notification.subject.type) {
-            SubjectType.PullRequest -> notification.detail?.let { detail ->
-                return when {
-                    detail.isPullRequestDraft() -> TYPE_TO_EMOJI["PullRequestDraft"]
-                    detail.isPullRequestClosed() -> TYPE_TO_EMOJI["PullRequestClosed"]
-                    detail.isPullRequestMerged() -> TYPE_TO_EMOJI["PullRequestMerged"]
-                    detail.isPullRequestOpen() -> TYPE_TO_EMOJI["PullRequestOpen"]
-                    else -> null
-                }
-            }
-            SubjectType.Issue -> notification.detail?.let { detail ->
-                return when {
-                    detail.isIssueClosed() -> TYPE_TO_EMOJI["IssueClosed"]
-                    detail.isIssueOpen() -> TYPE_TO_EMOJI["IssueOpen"]
-                    else -> null
-                }
-            }
-            SubjectType.Release -> TYPE_TO_EMOJI[type.name]
-            SubjectType.UNKNOWN -> null
-        }
-    }
-
-    private fun getIssueNumber(url: String): String = url.split("/").last()
 }